requirements.txt
numpy
pandas
scikit-learn
Flask
flask-cors
google-generativeai
gunicorn